pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "arbis-tools"
version = "0.1.0"
description = "Arbis Tools — Modular Secure Toolkit for Arch-based Systems"
authors = [
    { name="yberkut", email="yaroslav.berkut@gmail.com" }
]
dependencies = [
    "typer[all]",
    "pyyaml",
    "cryptography",
    "argon2-cffi"
]
readme = "README.md"
requires-python = ">=3.8"

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
arbis-ksm = "ksm.cli:app"